link pro inet funcs
